Initial version of sync-events plugin
This plugins allows to share stream events between two Gerrit instances
sharing the same git repositories and database.

The plugin needs to be installed in both instances and every time a
stream event occurs in one of the instances, the event is forwarded to
th other instance which re-plays it. This way, the output of the
stream-events command is the same, no matter what instance a client is
connected to.

Change-Id: I979ed4e21aa4c74de945b64242c3fee76b403861